Add envlist to tox settings
The envlist key in the tox configuration defines which environments
tox should create and execute by default when running the tox comma
nd. The environments in the envlist can also be run individually as
desired by using the -e argument to  the tox command and specifying
the environment name. To get started,  add  an  envlist  key to the
tox:tox section in your setup.cfg file with a value of py39

The next time you run tox, it will:
1. Create an isolated build of your package
2. Create a virtual environment with a copy of Python 3.9
3. Install your package in the virtual environment